moved non-commands to utils
