+++ cmd/cli/kubectl-kyverno/utils/common/common.go
@@ -21,8 +21,8 @@ import (
 	"github.com/go-logr/logr"
 	v1 "github.com/kyverno/kyverno/api/kyverno/v1"
 	report "github.com/kyverno/kyverno/api/policyreport/v1alpha2"
-	sanitizederror "github.com/kyverno/kyverno/cmd/cli/kubectl-kyverno/sanitizedError"
-	"github.com/kyverno/kyverno/cmd/cli/kubectl-kyverno/store"
+	sanitizederror "github.com/kyverno/kyverno/cmd/cli/kubectl-kyverno/utils/sanitizedError"
+	"github.com/kyverno/kyverno/cmd/cli/kubectl-kyverno/utils/store"
 	client "github.com/kyverno/kyverno/pkg/dclient"
 	"github.com/kyverno/kyverno/pkg/engine"
 	"github.com/kyverno/kyverno/pkg/engine/context"
