+++ cmd/cli/kubectl-kyverno/utils/common/fetch.go
